mcp data gen minimal/mcp-dataset-generator/requirements.txt
mcp-use>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
deepeval[openai]
matplotlib